python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
nltk==3.8.1
indic-nlp-library==0.92
rapidfuzz==3.5.2
jellyfish==0.9.0
symspellpy==6.7.7
SpeechRecognition==3.12.0
Sphinx==8.1.3